+++ devicemanagement/tests/TestDevMgmtApiSyncCmds.h
@@ -33,11 +33,21 @@ using TimeDuration = Clock::duration;
 #define DM_CMD_INVALID (256)
 #define INVALID_INPUT_SIZE (0)
 #define INVALID_OUTPUT_SIZE (0)
+#define ECID_LOT_ID_LENGTH 6
 
 DECLARE_bool(enable_trace_dump);
 DECLARE_string(trace_base_dir);
 DECLARE_string(trace_txt_dir);
 DECLARE_string(trace_bin_dir);
+
+typedef struct
+{
+    uint64_t lot_id;
+    uint8_t wafer_id;
+    uint8_t x_coordinate;
+    uint8_t y_coordinate;
+    char lot_id_str[ECID_LOT_ID_LENGTH + 1];
+} ecid_t;
 
 enum class Target { Silicon, Bemu, FullBoot, FullChip, SysEMU, Loopback };