+++ devicemanagement/tests/functional/FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds.cpp
@@ -53,8 +53,7 @@ TEST_F(FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds, getModuleUptime) {
 }
 
 TEST_F(FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds, getModulePower) {
-  // TODO: SW-13220: Enable back on Target::Silicon and SysEMU, following failure is seen with V2/V3 card
-  // Expected: (module_power->power) != (0), actual: '\0' vs 0
+  // TODO: Enable this test for silicon after SW-13282 is resolved.
   if (targetInList({Target::FullBoot, Target::FullChip, Target::Bemu})) {
     getModulePower(false /* Multiple devices */);
   } else {
@@ -88,19 +87,7 @@ TEST_F(FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds, setModuleActivePow
 }
 
 TEST_F(FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds, setThrottlePowerStatus) {
-  // TODO: SW-13220: Enable back on Target::Silicon, following failure is seen with V2/V3 card
-  // No SP trace event found!
-  // The txt trace file when failure occurs has no logged traces:
-  // bash-4.2$ cat devtrace/txt_files/dev0_traces.txt
-  //
-  // FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds.setThrottlePowerStatus
-  // -> SP Traces
-  if (getTestTarget() != Target::Silicon) {
-    setThrottlePowerStatus(false /* Multiple Devices */);
-  } else {
-    DV_LOG(INFO) << "Skipping the test since its not supported on current target";
-    FLAGS_enable_trace_dump = false;
-  }
+  setThrottlePowerStatus(false /* Multiple Devices */);
 }
 
 TEST_F(FunctionalTestDevMgmtApiThermalAndPowerMonitoringCmds, setModuleFrequency) {